Adds redis-diff command for columnular comparisons
./redis-diff command allows generation of diff tables between two redis
instances. The src db keyspace is enumerated using REDIS SCAN. That set
of keys is streamed to a series of workers (`-parallel N`) who query src
db and dest db for that key's value.

If `reflect.DeepCompare` doesn't return a matching result for two
values, then Discrepancy is sent to listening queue and written to
STDOUT in columnular format.